Change vault supply to shares
